Entrega

+++ src/profesor/profesor.service.ts
@@ -3,6 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { ProfesorEntity } from './profesor.entity';
 import { Repository } from 'typeorm';
 import { BusinessError, BusinessLogicException } from '../shared/business-errors';
+import { PropuestaEntity } from '../propuesta/propuesta.entity';
+import { PropuestaService } from '../propuesta/propuesta.service';
 
 @Injectable()
 export class ProfesorService {
@@ -12,31 +14,38 @@ export class ProfesorService {
     ){}
 
     async crearProfesor(profesor: ProfesorEntity): Promise<ProfesorEntity> {
+        const grupos = ["COMIT","IMAGINE","TICSW"]
+        if(!grupos.includes(profesor.grupoInvestigacion))
+          throw new BusinessLogicException("El grupo de investigacion no está en los esperados", BusinessError.BAD_REQUEST);
         return await this.profesorRepository.save(profesor);
     }
 
     async findProfesorById(id: string): Promise<ProfesorEntity> {
-        const profesor: ProfesorEntity = await this.profesorRepository.findOne({where: {id}, relations: ["artworks", "exhibitions"] } );
+        const profesor: ProfesorEntity = await this.profesorRepository.findOne({where: {id}, relations: ["propuestas"] } );
         if (!profesor)
-          throw new BusinessLogicException("The profesor with the given id was not found", BusinessError.NOT_FOUND);
+          throw new BusinessLogicException("El profesor con el id dado no existe", BusinessError.NOT_FOUND);
    
         return profesor;
     }
 
     async eliminarProfesor(id: string) {
         const profesor: ProfesorEntity = await this.profesorRepository.findOne({where:{id}});
-        if (!profesor)
-          throw new BusinessLogicException("The profesor with the given id was not found", BusinessError.NOT_FOUND);
-     
-        await this.profesorRepository.remove(profesor);
-    }
+            if (!profesor)
+              throw new BusinessLogicException("El profesor con el id dado no existe", BusinessError.NOT_FOUND);
+            
+            if (profesor.propuestas === null)
+              throw new BusinessLogicException("El profesor tiene propuestas con proyectos asociados", BusinessError.BAD_REQUEST);
+            await this.profesorRepository.remove(profesor);
+        }
 
-    async eliminarProfesorCedula(cedula: number) {
-        const profesor: ProfesorEntity = await this.profesorRepository.findOne({where:{cedula}});
-        if (!profesor)
-          throw new BusinessLogicException("The profesor with the given id was not found", BusinessError.NOT_FOUND);
-     
-        await this.profesorRepository.remove(profesor);
+        async eliminarProfesorCedula(cedula: number) {
+            const profesor: ProfesorEntity = await this.profesorRepository.findOne({where:{cedula}});
+            if (!profesor)
+              throw new BusinessLogicException("El profesor con el id dado no existe", BusinessError.NOT_FOUND);
+            
+            if (profesor.propuestas === null)
+              throw new BusinessLogicException("El profesor tiene propuestas con proyectos asociados", BusinessError.BAD_REQUEST);
+            await this.profesorRepository.remove(profesor);
     }
  
 }
